+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Add cumulative GIF output to `caustics render` through optional Flipbook.
+
 ## [0.1.0] - 2026-09-23
 
 - Initial release.